refactor(spec): split dataChecks into position, color, and temporal modules
Extract tier-2 scale compatibility from the monolithic dataChecks walk so
position config, x/y type checks, and color/fill checks can evolve independently.
Co-locate color data-aware validation tests. Public validate() is unchanged.
