+++ .changeset/spec-validate-data-checks-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split tier-2 dataChecks into position, color, and shared temporal modules, and co-locate color data-aware validation tests. Public validate() behavior is unchanged.
