+++ packages/spec/src/validate-data-checks.ts
@@ -1,32 +1,33 @@
 /**
  * Tier-2 data-aware walk — field existence, all-null columns, {stat} columns,
- * geom/stat type rules, and scale/type compatibility. Consumes a FieldEvidenceMap
- * built by validate-data-evidence.ts (or builds one when not pre-resolved).
+ * geom/stat type rules, then scale/type compatibility.
+ *
+ * Implementation:
+ *  - validate-data-checks-temporal.ts — shared temporalDecisionForField memoization
+ *  - validate-data-checks-position.ts — pre-evidence temporal axis config + x/y scale types
+ *  - validate-data-checks-color.ts — color/fill scale types + manual domain/range
+ *
+ * Consumes a FieldEvidenceMap built by validate-data-evidence.ts (or builds one
+ * when not pre-resolved).
  */
 import type { SpecError } from "./errors.js";
 import { didYouMean } from "./errors.js";
-import { configuredColorScaleType } from "./scale-helpers.js";
-import type { Aes, ChannelName, ColorScaleSpec, PositionScaleSpec } from "./schema.js";
-import { CHANNELS, GEOM_DEFAULTS, SEQUENTIAL_SCHEME_NAMES } from "./schema.js";
+import type { Aes, ChannelName } from "./schema.js";
+import { CHANNELS, GEOM_DEFAULTS } from "./schema.js";
 import type { ProfileFieldType, ValidateLimits, ValidateOptions } from "./validate-data.js";
-import { parseTemporalColumn, type TemporalDecision } from "./temporal-column.js";
-import {
-  canonicalTemporalParserKey,
-  temporalParserConfigurationError,
-  type TemporalParseOptions,
-  type TemporalParserSpec,
-} from "./temporal-parse.js";
-import {
-  parseTemporalInterval,
-  temporalLabelConfigurationError,
-  temporalLocaleConfigurationError,
-} from "./temporal-interval.js";
+import type { TemporalDecision } from "./temporal-column.js";
 import {
   effectiveChannel,
   resolveFieldEvidence,
-  type FieldEvidenceEntry,
   type ResolveFieldEvidenceResult,
 } from "./validate-data-evidence.js";
+import { checkColorScaleDataCompatibility } from "./validate-data-checks-color.js";
+import {
+  checkPositionScaleDataCompatibility,
+  scaleRequestsTime,
+  validateTemporalAxisConfiguration,
+} from "./validate-data-checks-position.js";
+import type { ChannelFieldUse } from "./validate-data-checks-temporal.js";
 
 /** Columns generated by each stat ({ stat } channel resolution contract).
  *  Documented per stat in its @ggsvelte/core module header. */
@@ -44,116 +45,6 @@ function isRecord(v: unknown): v is Record<string, unknown> {
   return typeof v === "object" && v !== null && !Array.isArray(v);
 }
 
-function appendTemporalKindMismatch(
-  errors: SpecError[],
-  input: {
-    axis: "x" | "y";
-    path: string;
-    field: string;
-    expected: "date" | "datetime" | undefined;
-    actual: "date" | "datetime" | null;
-  },
-): void {
-  const { axis, path, field, expected, actual } = input;
-  if (expected === undefined || actual === null || actual === expected) return;
-  errors.push({
-    code: "scale-type-mismatch",
-    path,
-    message: `scales.${axis}.temporalKind is "${expected}" but field "${field}" parses as ${actual}.`,
-    fix: {
-      description: `Use temporalKind "${actual}" or choose a parser that produces ${expected} values.`,
-    },
-  });
-}
-
-const AXIS_CHANNELS = ["x", "y"] as const;
-const COLOR_CHANNELS = ["color", "fill"] as const;
-const SEQUENTIAL_SCHEMES = new Set<string>(SEQUENTIAL_SCHEME_NAMES);
-
-/**
- * Resolve one field's temporal decision for scale compatibility checks.
- *
- * Memoized per dataChecks() call so multi-layer / multi-channel consumers of
- * the same field under the same parser+options pay O(n) once, not O(U·n).
- * When parser is auto and options match the evidence pass (no timezone /
- * disambiguation), reuse FieldEvidenceEntry.temporal instead of re-scanning.
- *
- * Cache key omits parseFailure and temporalKind — those only affect how the
- * decision is interpreted, not the decision itself.
- */
-function temporalDecisionForField(
-  cache: Map<string, TemporalDecision | null | undefined>,
-  field: string,
-  info: FieldEvidenceEntry | undefined,
-  parser: TemporalParserSpec | "auto",
-  options: TemporalParseOptions,
-): TemporalDecision | null | undefined {
-  const parserKey = parser === "auto" ? "auto" : canonicalTemporalParserKey(parser);
-  const key = `${field}\0${parserKey}\0${options.timezone ?? ""}\0${options.disambiguation ?? ""}`;
-  if (cache.has(key)) return cache.get(key);
-
-  let decision: TemporalDecision | null | undefined;
-  if (info?.values === null || info?.values === undefined) {
-    // Profile-backed fields: no inline values. temporal is always null today.
-    decision = info?.temporal ?? null;
-  } else if (
-    parser === "auto" &&
-    options.timezone === undefined &&
-    options.disambiguation === undefined &&
-    info.temporal !== null
-  ) {
-    // Evidence was built with inferTemporalColumn(column) — same auto defaults.
-    decision = info.temporal;
-  } else {
-    decision = parseTemporalColumn(info.values, parser, options).decision;
-  }
-  cache.set(key, decision);
-  return decision;
-}
-
-/** Type-preserving key for discrete domain identity (mirrors core encodeKey). */
-function discreteDomainKey(value: unknown): string {
-  switch (typeof value) {
-    case "string":
-      return value.startsWith("@") ? "@" + value : value;
-    case "number":
-      if (Number.isNaN(value)) return "@n:NaN";
-      if (Object.is(value, -0)) return "@n:-0";
-      return "@n:" + String(value);
-    case "boolean":
-      return "@b:" + String(value);
-    case "bigint":
-      return "@i:" + value.toString();
-    case "undefined":
-      return "@undefined";
-    default:
-      if (value === null) return "@null";
-      if (value instanceof Date) {
-        const t = value.getTime();
-        return "@d:" + (Number.isNaN(t) ? "NaN" : String(t));
-      }
-      return JSON.stringify(value);
-  }
-}
-
-function expectedManualDomainLength(
-  domain: unknown,
-  fieldValueLists: readonly (readonly unknown[] | null | undefined)[],
-): number | null {
-  if (Array.isArray(domain)) return domain.filter((value) => value !== null).length;
-  const seen = new Set<string>();
-  let sawValues = false;
-  for (const values of fieldValueLists) {
-    if (values === null || values === undefined) continue;
-    sawValues = true;
-    for (const value of values) {
-      if (value === null) continue;
-      seen.add(discreteDomainKey(value));
-    }
-  }
-  return sawValues ? seen.size : null;
-}
-
 export function dataChecks(
   spec: Record<string, unknown>,
   options: ValidateOptions,
@@ -162,83 +53,14 @@ export function dataChecks(
   preResolved?: ResolveFieldEvidenceResult,
 ): SpecError[] {
   const errors: SpecError[] = [];
+  // One cache per dataChecks call — shared by position and color checkers.
   const temporalDecisionCache = new Map<string, TemporalDecision | null | undefined>();
   const scales = isRecord(spec["scales"]) ? spec["scales"] : undefined;
-  const scaleRequestsTime = (axis: "x" | "y"): boolean => {
-    const config = scales?.[axis] as PositionScaleSpec | undefined;
-    return (
-      config?.type !== "band" &&
-      (config?.type === "time" ||
-        config?.parse !== undefined ||
-        config?.temporalKind !== undefined ||
-        config?.timezone !== undefined ||
-        config?.disambiguation !== undefined ||
-        config?.parseFailure !== undefined ||
-        config?.dateBreaks !== undefined ||
-        config?.dateMinorBreaks !== undefined ||
-        config?.dateLabels !== undefined ||
-        config?.locale !== undefined ||
-        config?.weekStart !== undefined)
-    );
-  };
-  const invalidTemporalAxes = new Set<"x" | "y">();
-  for (const axis of AXIS_CHANNELS) {
-    const config = scales?.[axis] as PositionScaleSpec | undefined;
-    const hasGuideTemporalOption =
-      config?.dateBreaks !== undefined ||
-      config?.dateMinorBreaks !== undefined ||
-      config?.dateLabels !== undefined ||
-      config?.locale !== undefined ||
-      config?.weekStart !== undefined;
-    if (
-      hasGuideTemporalOption &&
-      (config?.type === "band" || config?.type === "linear" || config?.type === "log")
-    ) {
-      invalidTemporalAxes.add(axis);
-      errors.push({
-        code: "scale-type-mismatch",
-        path: `/scales/${axis}`,
-        message: `scales.${axis} uses temporal break or label options with explicit type "${config.type}".`,
-        fix: {
-          description:
-            'Use type "time", a date/datetime scale helper, or remove the temporal option.',
-        },
-      });
-      continue;
-    }
-    if (config?.type === "band" || !scaleRequestsTime(axis)) continue;
-    let configurationError = temporalParserConfigurationError(config?.parse ?? "auto", {
-      ...(config?.timezone !== undefined && { timezone: config.timezone }),
-      ...(config?.disambiguation !== undefined && { disambiguation: config.disambiguation }),
-    });
-    if (configurationError === null) {
-      for (const interval of [config?.dateBreaks, config?.dateMinorBreaks]) {
-        if (interval === undefined) continue;
-        try {
-          parseTemporalInterval(interval);
-        } catch (error) {
-          configurationError = error instanceof Error ? error.message : "invalid temporal interval";
-          break;
-        }
-      }
-    }
-    if (configurationError === null && config?.dateLabels !== undefined) {
-      configurationError = temporalLabelConfigurationError(config.dateLabels);
-    }
-    if (configurationError === null && config?.locale !== undefined) {
-      configurationError = temporalLocaleConfigurationError(config.locale);
-    }
-    if (configurationError === null) continue;
-    invalidTemporalAxes.add(axis);
-    errors.push({
-      code: "scale-type-mismatch",
-      path: `/scales/${axis}`,
-      message: `scales.${axis} has invalid temporal configuration: ${configurationError}.`,
-      fix: {
-        description: "Correct the parser, interval, label, locale, or timezone configuration.",
-      },
-    });
-  }
+
+  // Pre-evidence: temporal axis configuration (errors even without data/profile).
+  const temporalConfig = validateTemporalAxisConfiguration(scales);
+  errors.push(...temporalConfig.errors);
+  const { invalidTemporalAxes } = temporalConfig;
 
   const resolved = preResolved ?? resolveFieldEvidence(spec, options, limits);
   if (resolved.status === "errors") return [...errors, ...resolved.errors];
@@ -255,15 +77,11 @@ export function dataChecks(
         : channel === "y" || channel === "ymin" || channel === "ymax"
           ? "y"
           : null;
-    return axis !== null && scaleRequestsTime(axis);
+    return axis !== null && scaleRequestsTime(scales, axis);
   };
 
-  interface AxisUse {
-    field: string;
-    path: string;
-  }
-  const axisFields: Record<"x" | "y", AxisUse[]> = { x: [], y: [] };
-  const colorFields: Record<"color" | "fill", AxisUse[]> = { color: [], fill: [] };
+  const axisFields: Record<"x" | "y", ChannelFieldUse[]> = { x: [], y: [] };
+  const colorFields: Record<"color" | "fill", ChannelFieldUse[]> = { color: [], fill: [] };
 
   for (let i = 0; i < layers.length; i++) {
     const layer = layers[i];
@@ -337,7 +155,7 @@ export function dataChecks(
         if (
           info !== null &&
           (info[1] === "nominal" || info[1] === "ordinal") &&
-          !scaleRequestsTime("y")
+          !scaleRequestsTime(scales, "y")
         ) {
           typeError(
             channel,
@@ -407,252 +225,22 @@ export function dataChecks(
     }
   }
 
-  // --- scale/type compatibility ----------------------------------------------
-  const typeOf = (field: string) => fields.get(field)?.type ?? null;
-
-  for (const axis of AXIS_CHANNELS) {
-    const config = scales?.[axis] as PositionScaleSpec | undefined;
-    const declared =
-      config?.type === "band" ? "band" : scaleRequestsTime(axis) ? "time" : config?.type;
-    if (declared === undefined || declared === "band" || invalidTemporalAxes.has(axis)) continue;
-    for (const use of axisFields[axis]) {
-      const type = typeOf(use.field);
-      if (type === null) continue;
-      if (declared === "time") {
-        const info = fields.get(use.field);
-        const decision = temporalDecisionForField(
-          temporalDecisionCache,
-          use.field,
-          info,
-          config?.parse ?? "auto",
-          {
-            ...(config?.timezone !== undefined && { timezone: config.timezone }),
-            ...(config?.disambiguation !== undefined && {
-              disambiguation: config.disambiguation,
-            }),
-          },
-        );
-        const censoredInvalid =
-          config?.parse !== undefined &&
-          config.parseFailure === "censor" &&
-          decision?.status === "invalid";
-        const profileTemporal = info?.values === null && type === "temporal";
-        if (decision?.status === "temporal" || censoredInvalid || profileTemporal) {
-          appendTemporalKindMismatch(errors, {
-            axis,
-            path: use.path,
-            field: use.field,
-            expected: config?.temporalKind,
-            actual: decision?.kind ?? null,
-          });
-          continue;
-        }
-        const firstFailure = decision?.failures?.[0];
-        const detail =
-          firstFailure === undefined
-            ? decision?.status === "ambiguous"
-              ? ` Automatic temporal inference was ambiguous between: ${decision.candidates.join(", ")}.`
-              : decision?.status === "invalid"
-                ? ` Automatic temporal inference failed whole-column validation for ${decision.failedCount} value(s).`
-                : ""
-            : ` Parser ${JSON.stringify(config?.parse ?? "auto")} rejected ${decision?.failedCount ?? 0} value(s), including row ${firstFailure.index}: ${JSON.stringify(firstFailure.value)}.`;
-        errors.push({
-          code: "scale-type-mismatch",
-          path: use.path,
-          message: `scales.${axis} requests time but field "${use.field}" is ${type}.${detail}`,
-          fix: {
-            description:
-              config?.parse === undefined
-                ? `Set scales.${axis}.parse to an explicit temporal order, or set scales.${axis}.type to "band" to keep categories.`
-                : `Correct the rejected values, choose the matching scales.${axis}.parse value, or use parseFailure: "censor" explicitly.`,
-          },
-        });
-        continue;
-      }
-      if (
-        (declared === "log" || declared === "linear") &&
-        (type === "nominal" || type === "ordinal")
-      ) {
-        errors.push({
-          code: "scale-type-mismatch",
-          path: use.path,
-          message: `scales.${axis}.type is "${declared}" but field "${use.field}" is ${type}; use a band scale (or a quantitative field).`,
-          fix: { description: `Set scales.${axis}.type to "band".` },
-        });
-      }
-    }
-  }
-
-  for (const channel of COLOR_CHANNELS) {
-    const config = scales?.[channel] as ColorScaleSpec | undefined;
-    const effectiveType = configuredColorScaleType(config);
-
-    // Manual range length vs domain / inferred unique values (union across layers).
-    if ((effectiveType === "manual" || config?.type === "manual") && Array.isArray(config?.range)) {
-      const range = config.range;
-      const expected = expectedManualDomainLength(
-        config.domain,
-        colorFields[channel].map((use) => fields.get(use.field)?.values),
-      );
-      if (expected !== null && range.length !== expected) {
-        errors.push({
-          code: "scale-manual-domain-range",
-          path: `/scales/${channel}`,
-          message: `The manual ${channel} scale needs one range color per domain value (${String(expected)} values, ${String(range.length)} colors).`,
-          fix: {
-            description: `Provide ${String(expected)} range colors, or set scales.${channel}.domain explicitly to match the range length.`,
-          },
-        });
-      }
-    }
-
-    const inferredFromSequentialScheme =
-      config?.type === undefined &&
-      config?.range === undefined &&
-      config?.scheme !== undefined &&
-      SEQUENTIAL_SCHEMES.has(config.scheme);
-    if (effectiveType !== "sequential" && effectiveType !== "binned") continue;
-    for (const use of colorFields[channel]) {
-      const type = typeOf(use.field);
-      if (
-        type === "temporal" &&
-        config?.transform !== undefined &&
-        config.transform !== "identity"
-      ) {
-        errors.push({
-          code: "scale-type-mismatch",
-          path: use.path,
-          message: `scales.${channel}.transform is "${config.transform}" but field "${use.field}" is temporal; temporal color scales permit only the identity transform.`,
-          fix: {
-            description: `Remove scales.${channel}.transform to keep temporal inference, or map a non-temporal quantitative field.`,
-          },
-        });
-        continue;
-      }
-
-      const requestsTemporal =
-        config?.temporalKind !== undefined ||
-        config?.parse !== undefined ||
-        config?.timezone !== undefined ||
-        config?.disambiguation !== undefined;
-
-      // Quantitative fields with temporal-only options fail at resolve time.
-      if (type === "quantitative" && requestsTemporal) {
-        const info = fields.get(use.field);
-        const decision = temporalDecisionForField(
-          temporalDecisionCache,
-          use.field,
-          info,
-          config?.parse ?? "auto",
-          {
-            ...(config?.timezone !== undefined && { timezone: config.timezone }),
-            ...(config?.disambiguation !== undefined && {
-              disambiguation: config.disambiguation,
-            }),
-          },
-        );
-        if (decision?.status !== "temporal") {
-          errors.push({
-            code: "scale-type-mismatch",
-            path: use.path,
-            message: `scales.${channel} requests temporal colors but field "${use.field}" is quantitative (numbers are not treated as temporal without a successful epoch parse).`,
-            fix: {
-              description: `Map a temporal field, use a working parse: { epoch: "ms" | "s" }, or remove temporal color options.`,
-            },
-          });
-          continue;
-        }
-        if (
-          config?.temporalKind !== undefined &&
-          decision.kind !== null &&
-          decision.kind !== undefined &&
-          decision.kind !== config.temporalKind
-        ) {
-          errors.push({
-            code: "scale-type-mismatch",
-            path: use.path,
-            message: `scales.${channel} requests temporal kind "${config.temporalKind}" but field "${use.field}" parses as "${decision.kind}".`,
-            fix: {
-              description: `Use the ${decision.kind ?? "matching"} color helper or correct the source precision.`,
-            },
-          });
-          continue;
-        }
-      }
-
-      if (type !== "nominal" && type !== "ordinal") continue;
-
-      if (!requestsTemporal) {
-        errors.push({
-          code: "scale-type-mismatch",
-          path: use.path,
-          message: inferredFromSequentialScheme
-            ? `scales.${channel}.scheme is "${config.scheme}" and selects a sequential scale, but field "${use.field}" is ${type}; sequential color ramps need quantitative values.`
-            : `scales.${channel}.type is "${effectiveType}" but field "${use.field}" is ${type}; quantitative color scales need quantitative or temporal values.`,
-          fix: inferredFromSequentialScheme
-            ? {
-                description: `Set scales.${channel}.scheme to a categorical scheme, remove it to infer an ordinal scale from "${use.field}", or map a quantitative field.`,
-                example: { scheme: "observable10" },
-              }
-            : { description: `Set scales.${channel}.type to "ordinal".` },
-        });
-        continue;
-      }
-
-      const info = fields.get(use.field);
-      const decision = temporalDecisionForField(
-        temporalDecisionCache,
-        use.field,
-        info,
-        config?.parse ?? "auto",
-        {
-          ...(config?.timezone !== undefined && { timezone: config.timezone }),
-          ...(config?.disambiguation !== undefined && {
-            disambiguation: config.disambiguation,
-          }),
-        },
-      );
-      const censoredInvalid =
-        config?.parse !== undefined &&
-        config.parseFailure === "censor" &&
-        decision?.status === "invalid";
-      if (decision?.status === "temporal" || censoredInvalid) {
-        if (
-          decision?.status === "temporal" &&
-          config?.temporalKind !== undefined &&
-          decision.kind !== config.temporalKind
-        ) {
-          errors.push({
-            code: "scale-type-mismatch",
-            path: use.path,
-            message: `scales.${channel} requests temporal kind "${config.temporalKind}" but field "${use.field}" parses as "${decision.kind}".`,
-            fix: {
-              description: `Use the ${decision.kind ?? "matching"} color helper or correct the source precision.`,
-            },
-          });
-        }
-        continue;
-      }
-
-      const detail =
-        decision?.status === "ambiguous"
-          ? ` Automatic temporal inference was ambiguous between: ${decision.candidates.join(", ")}.`
-          : decision?.status === "invalid"
-            ? ` Temporal parsing rejected ${String(decision.failedCount)} value(s).`
-            : "";
-      errors.push({
-        code: "scale-type-mismatch",
-        path: use.path,
-        message: `scales.${channel} requests temporal colors but field "${use.field}" is ${type}.${detail}`,
-        fix: {
-          description:
-            config?.parse === undefined
-              ? `Set scales.${channel}.parse to the exact temporal order, or use type: "ordinal".`
-              : `Correct the rejected values, choose the matching parser, or set parseFailure: "censor" explicitly.`,
-        },
-      });
-    }
-  }
+  // --- scale/type compatibility (order preserved for diagnostics) ------------
+  errors.push(
+    ...checkPositionScaleDataCompatibility({
+      scales,
+      fields,
+      axisFields,
+      invalidTemporalAxes,
+      temporalDecisionCache,
+    }),
+    ...checkColorScaleDataCompatibility({
+      scales,
+      fields,
+      colorFields,
+      temporalDecisionCache,
+    }),
+  );
 
   return errors;
 }
